Add the client instance to the global scope
