+++ lib/db_connect.go
@@ -3,41 +3,24 @@ package lib
 import (
 	"context"
 	"fmt"
-	"log"
-	"os"
 
-	"github.com/joho/godotenv"
+	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 func ConnectHello() {
 	fmt.Println("Hello, this is the connect package")
 }
 
-// SetupConnection creates a connection to the mongodb database given the uri setup in the environment variable
-// If an error would occur in the .env file or while loading the variable the function will log the error
-// If an error would occur with the client connection, the function will close the connection and panic
-// The returned value is a Client struct which can be used to handle the connection to the database
-func SetupConnection() mongo.Client {
+// TestConnection sends a ping to the database via the client instance
+// It panics if the connection is down, otherwise prints a confirmation
+func TestConnection(client mongo.Client) {
 
-	if err := godotenv.Load(); err != nil {
-		log.Println("No .env file found")
-	}
-	uri := os.Getenv("MONGODB_URI")
-	if uri == "" {
-		log.Fatal("You must set your 'MONGODB_URI' environment variable. See\n\t https://www.mongodb.com/docs/drivers/go/current/usage-examples/#environment-variable")
-	}
-
-	client, err := mongo.Connect(context.TODO(), options.Client().ApplyURI(uri))
-	if err != nil {
+	// Send a ping to confirm a successful connection
+	var result bson.M
+	if err := client.Database("admin").RunCommand(context.TODO(), bson.D{{"ping", 1}}).Decode(&result); err != nil {
 		panic(err)
 	}
-	defer func() {
-		if err := client.Disconnect(context.TODO()); err != nil {
-			panic(err)
-		}
-	}()
+	fmt.Println("Pinged your deployment. You successfully connected to MongoDB!")
 
-	return *client
 }
